+++ ree_lib/lib/ree_lib/packages/ree_decorators/package/ree_decorators/decorators/throws.rb
@@ -1,0 +1,11 @@
+# frozen_string_literal: true
+
+class ReeDecorators::Throws
+  include ReeDecorators::DSL
+
+  decorator :throws
+
+  def build_context(*exception_classes)
+    exception_classes
+  end
+end