+++ ree/CHANGELOG.md
@@ -1,5 +1,7 @@
 ## [Unreleased]
 
+- Added `on_link` hook to `ObjectDsl`
+
 ## [1.1.2] - 2025-05-14
 
 - Forward compatibility fixes for ruby 3.4