+++ ree_lib/CHANGELOG.md
@@ -1,5 +1,7 @@
 ## [Unreleased]
 
+- Added `ReeDecorators::DSL` module for method decoration functionality
+
 ## [1.3.0] - 2025-05-14
 
 - Forward compatibility fixes for ruby 3.4